+++ Example/AppDelegate.swift
@@ -10,18 +10,3 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         return true
     }
 }
-
-extension UIApplication {
-  class func topViewController(controller: UIViewController? = UIApplication.shared.keyWindow?.rootViewController) -> UIViewController? {
-    if let tabController = controller as? UITabBarController {
-      return topViewController(controller: tabController.selectedViewController)
-    }
-    if let navController = controller as? UINavigationController {
-      return topViewController(controller: navController.visibleViewController)
-    }
-    if let presented = controller?.presentedViewController {
-      return topViewController(controller: presented)
-    }
-    return controller
-  }
-}
