+++ Example/en.lproj/Localizable.strings
@@ -21,7 +21,11 @@
 "wait_message" = "Please wait...";
 "ok" = "OK";
 "cancel" = "Cancel";
+"yes" = "Yes";
 
 "test_mode_publishing" = "Published";
 "test_mode_testing" = "Testing";
 "test_mode_default" = "Default";
+
+"title_allow" = "Allow";
+"title_dont_allow" = "Don't Allow";
